feat: add dotenv and move server files to src
