+++ apps/api/.example.env
@@ -1,0 +1,2 @@
+DATABASE_URL=
+PORT=
