+++ apps/api/package.json
@@ -5,10 +5,12 @@
   "license": "MIT",
   "scripts": {
     "build": "npx tsc",
-    "dev": "nodemon app.ts"
+    "dev": "nodemon ./src/app.ts"
   },
   "dependencies": {
     "@prisma/client": "^4.5.0",
+    "dotenv": "^16.0.3",
+    "envalid": "^7.3.1",
     "express": "^4.18.2"
   },
   "devDependencies": {
