+++ apps/api/src/app.ts
@@ -1,10 +1,12 @@
 import express from 'express';
 import { router as exams } from './routes/exams';
+import { env } from './config';
 
 const app = express();
 app.use(express.json());
 
 app.use('/api/v1/exams', exams);
 
-const PORT = 3000;
-app.listen(PORT, () => console.log(`Server is listening on port ${PORT}`));
+app.listen(env.PORT, () =>
+  console.log(`🚀 Server ready at http://localhost:${env.PORT}`)
+);
