+++ apps/api/src/config.ts
@@ -1,0 +1,9 @@
+import { cleanEnv, port, str } from 'envalid';
+import { default as dotenv } from 'dotenv';
+
+dotenv.config();
+
+export const env = cleanEnv(process.env, {
+  PORT: port(),
+  DATABASE_URL: str(),
+});
