+++ apps/api/src/controllers/exams.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import prisma from 'prisma';
+import prisma from '../prisma';
 
 export const getAllExams = async (req: Request, res: Response) => {
   try {
