+++ apps/api/tsconfig.json
@@ -1,12 +1,9 @@
 {
   "compilerOptions": {
-    // "target": "es2016",
-    // "module": "commonjs",
     "sourceMap": true,
     "outDir": "dist",
     "esModuleInterop": true,
     "lib": ["ESNext"],
-    "strict": true,
-    "baseUrl": "."
+    "strict": true
   }
 }
